+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the "regex-context" extension will be documented in this file.
 
+## [0.0.5] -- 2022-01-19 22:03:42
+
+- Improve text around the `isActive` checkbox
+- Add a command to toggle whether or not it's active -- show the view also
+- Planned
+  - Add a find mode which takes the results into a new document
+
 ## [0.0.4]
 
 - Add initial support for nested contexts